Move SyncValidator to validation package and split into LoginValidator and SyncValidator
